Upstream's late-August work: type and naming corrections throughout, the
split of several oversized translation units, and decompilation fixes.
219 files conflicted; each was resolved by taking both sides -- upstream's
renames and refactors adopted, the port's host fixes carried onto them --
rather than letting either side win by rule, which is what made the first
attempt at this merge unbuildable.

Four things went wrong in ways worth recording, because each was silent:

- lbFile_800163D8 was renamed lbFileGetSize upstream, and the two callers
  in grdatfiles.c that still used the old name linked against a weak stub
  that returns zero. Every stage archive then loaded as zero bytes, so the
  stage never converted and a match drew its HUD over an empty scene at a
  seventh of the usual draw count. The stub is deleted so the next rename
  is a link error.

- The character select read its scene table in place again through
  upstream's spelling rather than the port's converting accessor. Four
  pointers at GameCube offsets 0/4/8/0xC are at 0/8/0x10/0x18 here, so the
  camera descriptor came out as two spliced offsets and the scene crashed
  on entry.

- Upstream decompiled the real refraction descriptors, so a code path that
  had never run on the host began to: it reads raw big-endian DAT bytes as
  a host structure and crashed entering a versus match. Refraction stays
  off here, as it has been all along, until that descriptor is converted.

- src/Runtime/platform.h typedefs ssize_t as int, which is a hard conflict
  with the host's LP64 definition. Android needed one more: Bionic has no
  __jmp_buf typedef where glibc does.

Also: the PC stub files no longer define 57 symbols that upstream now
defines for real; the sislib split files are not built, because the port's
own sislib.c still holds every function they define, with the host fixes
they lack; and the renamed scene entry points keep their stubs under the
new names.

Golden suite: every case within baseline except match_move:315, which is
0.7/255 worse. Android builds, and a versus match runs at 60 fps on the
tablet with no crash.

+++ src/melee/ft/chara/ftLuigi/ftLg_SpecialLw.c
@@ -4,7 +4,6 @@
 
 #include <platform.h>
 
-#include "ef/eflib.h"
 #include "ef/efsync.h"
 
 #include "forward.h"
@@ -21,7 +20,6 @@
 
 #include "ftCommon/ftCo_Fall.h"
 #include "ftCommon/ftCo_FallSpecial.h"
-#include "lb/lbrefract.h"
 
 #include <dolphin/mtx.h>
 
@@ -45,8 +43,8 @@ static inline void ftLuigi_SpecialLw_SetVars(HSD_GObj* gobj)
     ftLuigiAttributes* luigiAttrs = fp->dat_attrs;
     fp->cmd_vars[0] = 0;
     fp->cmd_vars[1] = 0;
-    fp->mv.lg.SpecialLw.groundVelX = (float) 0.0f;
-    fp->mv.lg.SpecialLw.unk = (s32) luigiAttrs->x88_LUIGI_CYCLONE_UNK + 1;
+    fp->mv.lg.SpecialLw.groundVelX = 0.0f;
+    fp->mv.lg.SpecialLw.unk = luigiAttrs->x88_LUIGI_CYCLONE_UNK + 1;
     fp->mv.lg.SpecialLw.isUnkColl = false;
 }
 
@@ -84,8 +82,8 @@ void ftLg_SpecialLw_Enter(HSD_GObj* gobj)
     Fighter_ChangeMotionState(gobj, ftLg_MS_SpecialAirLw, 0, 0.0f, 1.0f, 0.0f,
                               NULL);
     ftAnim_8006EBA4(gobj);
-    fp2->self_vel.y = (float) (luigiAttrs->x70_LUIGI_CYCLONE_TAP_MOMENTUM -
-                               luigiAttrs->x8C_LUIGI_CYCLONE_TAP_Y_VEL_MAX);
+    fp2->self_vel.y = (luigiAttrs->x70_LUIGI_CYCLONE_TAP_MOMENTUM -
+                       luigiAttrs->x8C_LUIGI_CYCLONE_TAP_Y_VEL_MAX);
     ftCommon_ClampSelfVelX(fp, luigiAttrs->x78_LUIGI_CYCLONE_MOMENTUM_X_AIR);
     ftLuigi_SpecialLw_SetVars(gobj);
     ftLuigi_SpecialLw_SetCall(gobj);
@@ -119,7 +117,7 @@ void ftLg_SpecialAirLw_Enter(HSD_GObj* gobj)
         cycloneVar = luigiAttrs->x8C_LUIGI_CYCLONE_TAP_Y_VEL_MAX;
     }
     fp2->self_vel.y =
-        (float) (luigiAttrs->x70_LUIGI_CYCLONE_TAP_MOMENTUM - cycloneVar);
+        (luigiAttrs->x70_LUIGI_CYCLONE_TAP_MOMENTUM - cycloneVar);
     ftCommon_ClampSelfVelX(fp, luigiAttrs->x78_LUIGI_CYCLONE_MOMENTUM_X_AIR);
     ftLuigi_SpecialLw_SetVars(gobj);
     ftLuigi_SpecialLw_SetCall(gobj);
@@ -270,7 +268,7 @@ static ftCollisionBox ftLg_SpecialLw_CollisionBox = {
 static inline void ftLuigi_SpecialLw_UnkAngle(HSD_GObj* gobj)
 {
     Fighter* fp = GET_FIGHTER(gobj);
-    if (((u32) fp->cmd_vars[3] != 0U) &&
+    if ((fp->cmd_vars[3] != 0U) &&
         ((s32) fp->mv.lg.SpecialLw.isUnkColl != false))
     {
         ftPartSetRotX(fp, 0,
